Add referral code generation and update migrations for CustomUser model
- Implemented a management command to generate referral codes for users without them.
- Updated the CustomUser model to include 'has_first_deposit' and 'referral_code' fields.
- Altered the 'roi' field in InvestmentPlan to be a FloatField.
- Added a ForeignKey field 'referred_by' to the CustomUser model for tracking referrals.
- Adjusted the max_length of 'referral_code' to accommodate longer codes.

+++ administration/models.py
@@ -14,7 +14,7 @@ class CryptoChannel(models.Model):
 class InvestmentPlan(models.Model):
     name = models.CharField(max_length=30, unique=True)
     label = models.CharField(max_length=30, unique=True, blank=True, null=True)
-    roi = models.IntegerField()
+    roi = models.FloatField()
     icon= models.ImageField(upload_to='icon')
     
     def __str__(self):
